+++ CHANGES.rst
@@ -25,6 +25,7 @@ Bug fix:
   * MXKRoomViewController: Add viaServers parameter to joinRoomWithRoomIdOrAlias.
   * MXKAccount: Make setEnablePushKitNotifications async.
   * MXKAccount: Rename enablePushKitNotifications to hasPusherForPushKitNotifications.
+  * MXKAccount: Remove deletePushKitPusher. Use setEnablePushKitNotifications:NO instead.
 
 Changes in MatrixKit in 0.9.9 (2019-05-03)
 ==========================================